// apps/web/src/entities/generated/entity-details.gen.ts
export const detailEntities = [
  "product",
  "recipe",
  "ingredient",
  "location",
  "inventory",
  "meal",
  "ledgerParty",
  "ledgerTransfer",
  "project",
  "task",
  "vendor",
  "purchase",
  "financialAccount",
  "financialTransaction",
  "wish",
  "expense",
] as const;
export type DetailEntity = (typeof detailEntities)[number];

export type EntityDetailByEntity = {
  product: z.output<typeof productWithFoodOut>;
  recipe: z.output<typeof recipeOut>;
  ingredient: z.output<typeof ingredientWithFoodOut>;
  location: z.output<typeof infLocation>;
  inventory: z.output<typeof inventoryWithLocationAndProductOut>;
  meal: z.output<typeof mealOut>;
  ledgerParty: z.output<typeof ledgerPartyOut>;
  ledgerTransfer: z.output<typeof ledgerTransferOut>;
  project: z.output<typeof projectOut>;
  task: z.output<typeof taskOut>;
  vendor: z.output<typeof vendorOut>;
  purchase: z.output<typeof purchaseOut>;
  financialAccount: z.output<typeof financialAccountOut>;
  financialTransaction: z.output<typeof financialTransactionOut>;
  wish: z.output<typeof wishOut>;
  expense: z.output<typeof expenseOut>;
};

export type EntityDetailInputByEntity = {
  product: { entity: "product"; shortcode: z.input<typeof productShortcode> };
  recipe: { entity: "recipe"; shortcode: z.input<typeof recipeShortcode> };
  ingredient: {
    entity: "ingredient";
    shortcode: z.input<typeof ingredientShortcode>;
  };
  location: {
    entity: "location";
    shortcode: z.input<typeof locationShortcode>;
  };
  inventory: {
    entity: "inventory";
    shortcode: z.input<typeof inventoryShortcode>;
  };
  meal: { entity: "meal"; shortcode: z.input<typeof mealShortcode> };
  ledgerParty: {
    entity: "ledgerParty";
    shortcode: z.input<typeof ledgerPartyShortcode>;
  };
  ledgerTransfer: {
    entity: "ledgerTransfer";
    shortcode: z.input<typeof ledgerTransferShortcode>;
  };
  project: { entity: "project"; shortcode: z.input<typeof projectShortcode> };
  task: { entity: "task"; shortcode: z.input<typeof taskShortcode> };
  vendor: { entity: "vendor"; shortcode: z.input<typeof vendorShortcode> };
  purchase: {
    entity: "purchase";
    shortcode: z.input<typeof purchaseShortcode>;
  };
  financialAccount: {
    entity: "financialAccount";
    shortcode: z.input<typeof financialAccountShortcode>;
  };
  financialTransaction: {
    entity: "financialTransaction";
    shortcode: z.input<typeof financialTransactionShortcode>;
  };
  wish: { entity: "wish"; shortcode: z.input<typeof wishShortcode> };
  expense: { entity: "expense"; shortcode: z.input<typeof expenseShortcode> };
};